+++ patterns/index.styl
@@ -49,4 +49,4 @@
 
 /* Add project specific styles here */
 @import './HenrysPerformanceComponent/henrysPerformanceComponent.styl'
-@import './HenrysSEOTeaserGridModul/henrysSEOTeaserGridModul.styl'
+@import './HenrysUndRonsSEOTeaserGridModul/henrysUndRonsSEOTeaserGridModul.styl'
